pyproject.toml
[tool.poetry]
name = "repl_python3_BotTaverneira"
version = "0.1.0"
description = ""
authors = ["RitaSantosRg <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
youtube-dl = "^2021.4.26"
discord = "^1.0.1"
flask = "^2.0.0"
python-dotenv = "^0.17.1"
PyNaCl = "^1.4.0"
"discord.py" = "^1.7.2"
replit = "^3.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
